refactor model - type adjust

+++ app/order_api/domain/model/item.go
@@ -2,5 +2,5 @@ package model
 
 type Item struct {
 	ItemID ItemID
-	Price  int64
+	Price  int
 }
